added functionality, fixed probability
Added the functionality to print the probabilitie of the next words of
one word: getTransitions(Word).

Fixed the weird probabilities. They are not longer “hacked” and are
shown as the real probability.

The random-generation now follows real randomness and isn’t hacked
anymore.

added some comments.

added possible logging  -  to see how the program chooses words.